refactor(app-ai, app-media): decompose AiUsagePage and WatchlistPage
Split AI observability UI into focused modules under pages/ai-usage and
watchlist UI into pages/watchlist (hook, DnD desktop, mobile list, item).

Closes #2078.
