+++ packages/app-ai/src/pages/ai-usage/types.ts
@@ -1,0 +1,9 @@
+export interface HistoryRecord {
+  date: string;
+  calls: number;
+  inputTokens: number;
+  outputTokens: number;
+  costUsd: number;
+  cacheHits: number;
+  errors: number;
+}
